Implement MQ client utils to resolve circular dependency with neon_utils (#77)
* Added infinite restart support (#69)


* Increment Version

(cherry picked from commit 695bc80ab14bcd571a7b75196aba1d8f61d8515a)

* Increment Version

* Increment version to 0.6.0 (#71)

* Implement MQ client utils to resolve circular dependency with neon_utils

* Update util test case

* Specify `/test` vhost for unit tests
Update MQConnector docstring

* Troubleshooting unit test failures

* revert test troubleshooting changes

---------

+++ neon_mq_connector/utils/connection_utils.py
@@ -53,7 +53,7 @@ def get_timeout(backoff_factor: float, number_of_retries: int) -> float:
 
 
 def retry(callback_on_exceeded: Union[str, Callable] = None, callback_on_attempt_failure: Union[str, Callable] = None,
-          num_retries: int = 3, backoff_factor: int = 5, use_self: bool = False,
+          num_retries: int = 3, backoff_factor: float = 5, use_self: bool = False,
           callback_on_attempt_failure_args: list = None, callback_on_exceeded_args: list = None):
     """
         Decorator for generic retrying function execution
